requirements.txt
feedparser
google-genai
google-cloud-firestore